pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.2"
uvicorn = {extras = ["standard"], version = "^0.27.0.post1"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
sqlalchemy = "^2.0.25"
pydantic-settings = "^2.1.0"
pydantic = {extras = ["email"], version = "^2.6.1"}
psycopg2-binary = "^2.9.9"
python-multipart = "^0.0.7"
bcrypt = "^4.1.2"
langchain-google-genai = "^0.0.9"
langchain = "^0.1.8"
pypdf2 = "^3.0.1"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
black = "^24.1.1"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"